+++ src/Resources/app/administration/src/module/sidworks-search-results/page/sidworks-search-results-list/index.js
@@ -7,15 +7,35 @@ export default {
 
     inject: [
         'acl',
-        'repositoryFactory'
+        'repositoryFactory',
+        'filterFactory'
     ],
 
     data() {
         return {
-            searchResultsRepository: null,
             searchResultsRepositoryItems: null,
-            total: 0
+            total: 0,
+            page: 1,
+            limit: 25,
+            storeKey: 'grid.filter.searchResults',
+            defaultFilters: [
+                'search-result-filter',
+                'times-searched-filter',
+                'results-count-filter',
+                'sales-channel-filter'
+            ],
+            activeFilterNumber: 0,
+            filterCriteria: []
         };
+    },
+
+    watch: {
+        defaultCriteria: {
+            handler() {
+                this.getList();
+            },
+            deep: true,
+        },
     },
 
     metaInfo() {
@@ -25,21 +45,45 @@ export default {
     },
 
     methods: {
-        getList() {
-            const criteria = new Criteria(1, 25);
-            criteria.addSorting(Criteria.sort('timesSearched', 'DESC'));
-            criteria.addAssociation('salesChannel');
+        async getList() {
+            this.isLoading = true;
 
-            this.searchResultsRepository
-                .search(criteria, Shopware.Context.api)
-                .then((result) => {
-                    this.searchResultsRepositoryItems = result;
-                    this.total = result.total;
-                });
+            const criteria = await Shopware.Service('filterService')
+                .mergeWithStoredFilters(this.storeKey, this.defaultCriteria);
+
+            this.activeFilterNumber = criteria.filters.length;
+
+            try {
+                const items = await this.searchResultsRepository.search(criteria);
+                this.total = items.total;
+                this.searchResultsRepositoryItems = items;
+                this.isLoading = false;
+                this.selection = {};
+            } catch {
+                this.isLoading = false;
+            }
+        },
+        updateCriteria(criteria) {
+            this.page = 1;
+            this.filterCriteria = criteria;
         }
     },
 
     computed: {
+        defaultCriteria() {
+            const defaultCriteria = new Criteria(this.page, this.limit);
+            defaultCriteria.addSorting(Criteria.sort('timesSearched', 'DESC'));
+            defaultCriteria.addAssociation('salesChannel');
+
+            this.filterCriteria.forEach(filter => {
+                defaultCriteria.addFilter(filter);
+            });
+
+            return defaultCriteria;
+        },
+        searchResultsRepository() {
+            return this.repositoryFactory.create('sidworks_search_results');
+        },
         columns() {
             return [
                 {
@@ -63,11 +107,46 @@ export default {
                     label: this.$t('sidworks-search-results.list.salesChannel')
                 }
             ];
-        }
+        },
+        listFilters() {
+            return this.filterFactory.create('sidworks_search_results', {
+                'search-result-filter': {
+                    property: 'searchTerm',
+                    type: 'string-filter',
+                    criteriaFilterType: 'contains',
+                    label: this.$tc('sidworks-search-results.list.searchTerm'),
+                    placeholder: this.$tc('sidworks-search-results.list.searchTerm'),
+                    valueProperty: 'key',
+                    labelProperty: 'key'
+                },
+                'times-searched-filter': {
+                    property: 'timesSearched',
+                    type: 'number-filter',
+                    label: this.$tc('sidworks-search-results.list.timesSearched'),
+                    fromFieldLabel: null,
+                    toFieldLabel: null,
+                    fromPlaceholder: this.$tc('global.default.from'),
+                    toPlaceholder: this.$tc('global.default.to'),
+                },
+                'results-count-filter': {
+                    property: 'resultsCount',
+                    type: 'number-filter',
+                    label: this.$tc('sidworks-search-results.list.resultsCount'),
+                    fromFieldLabel: null,
+                    toFieldLabel: null,
+                    fromPlaceholder: this.$tc('global.default.from'),
+                    toPlaceholder: this.$tc('global.default.to'),
+                },
+                'sales-channel-filter': {
+                    property: 'salesChannel',
+                    label: this.$tc('sidworks-search-results.list.salesChannel'),
+                    placeholder: this.$tc('sidworks-search-results.list.salesChannel'),
+                },
+            });
+        },
     },
 
     created() {
-        this.searchResultsRepository = this.repositoryFactory.create('sidworks_search_results');
         this.getList();
     }
 };